chore: Change to TS rollup.config. docs: Add keywords and improve description.
