+++ tsconfig.json
@@ -1,21 +1,21 @@
 {
   "compilerOptions": {
-    "noImplicitAny": true,
     "noEmitOnError": false,
     "target": "ES2020",
     "module": "ESNext",
     "moduleResolution": "Node",
     "esModuleInterop": true,
     "strict": true,
-    "strictNullChecks": true,
-    "alwaysStrict": true,
     "outDir": "build",
     "declaration": true,
     "emitDeclarationOnly": true,
-    "sourceMap": true
+    "sourceMap": true,
+    "skipLibCheck": true,
+    // "exactOptionalPropertyTypes": true,
+    "forceConsistentCasingInFileNames": true
   },
-  "compileOnSave": true,
   "include": [
-    "src"
+    "src",
+    "rollup.config.ts"
   ]
 }
